requirements.txt
## inserire qui librerie esterne utilizzate.
time
json
random
paho.mqtt.client
pymongo
datetime